refactor: separate state to different file
