+++ src/main.rs
@@ -92,6 +92,7 @@ fn get_matches() -> clap::ArgMatches<'static> {
         (@arg DEFAULT_WIDGET_COUNT: --default_widget_count +takes_value "Which number of the selected widget type to select, from left to right, top to bottom.  Defaults to 1.")
         (@arg USE_OLD_NETWORK_LEGEND: --use_old_network_legend "Use the older (pre-0.4) network widget legend.")
         (@arg HIDE_TABLE_GAP: --hide_table_gap "Hides the spacing between the table headers and entries.")
+        (@arg BATTERY: --battery "Shows the battery widget in default or basic mode.  No effect on custom layouts.")
 	)
         .get_matches()
 }
